Add ReadOrPeekWrapper to unify operand loading
Loading operands is complex and full of little nits that need careful
testing. But there need to be two implementations of the same logic,
one that loads operands for disassembly, using only peeks and immutable
access to the cpu.
And another one used during execution that will be able to mutate
the cpu.
ReadOrPeekWrapper is a bit of ugly boilerplate to make sure we
only need one implementation of the tricky logic.

+++ sres_emulator/src/memory.rs
@@ -2,7 +2,6 @@ use std::fmt::Display;
 use std::fmt::Formatter;
 use std::ops::Add;
 
-use crate::uint::RegisterSize;
 use crate::uint::U16Ext;
 use crate::uint::U32Ext;
 use crate::uint::UInt;
@@ -58,7 +57,7 @@ impl Add<u16> for Address {
     #[inline]
     fn add(self, rhs: u16) -> Self::Output {
         let bank = self.bank;
-        let offset = self.offset.wrapping_add(rhs as u16);
+        let offset = self.offset.wrapping_add(rhs);
         Address { bank, offset }
     }
 }
